assignment given on 12/03/2026

+++ 12-03-2026/project_6/project_6/src/main/java/com/example/project_6/repository/TeachersRepo.java
@@ -1,9 +1,8 @@
 package com.example.project_6.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import com.example.project_6.entity.TeachersEntity;
 
-public interface TeachersRepo extends JpaRepository<TeachersEntity, Integer> {
+public interface TeachersRepo extends JpaRepository<TeachersEntity, Long> {
     
 }
